+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 3.5.0
 ### Added
 - Added `hidden` parameter to `CliktCommand`, which will prevent the command from being displayed as a subcommand in help output  ([#353](https://github.com/ajalt/clikt/issues/353))
 - Publish artifacts for the `macosArm64` target. Note that this target is not tested on CI. ([#352](https://github.com/ajalt/clikt/issues/352))